pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "git-commits-graph"
version = "0.1.0"
description = "Display graph of changes in number of lines in project or changed lines"
repository = "https://github.com/izikeros/git-commits-graph"
authors = ["Krystian Safjan <ksafjan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
git-commits-graph = 'git_commits_graph.main:main'

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.27"
click = "^8.1.3"
pandas = "^1.4.2"
matplotlib = "^3.5.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pre-commit = "^2.19.0"
tox = "^3.25.0"
tox-poetry = "^0.4.1"
plotly = "^5.8.0"

[tool.poetry.group.dev.dependencies]
tox = "^3.27.1"
